pyproject.toml
[project]
name = "git-bard"
version = "1.0.0"
description = "A CLI tool to generate git commit messages using AI"
readme = "README.md"
authors = [
    { name = "仌 POURYA", email = "ogpourya@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "google-genai",
]

[project.scripts]
git-bard = "git_bard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["git_bard.py"]
